Revert "feat: refactor best/worst record api in response time (#323)"
This reverts commit 64821597b0a1ac7e2fb180ec546a4c0094b00125.
