+++ src/caret_analyze/record/records_service/response_time.py
@@ -257,12 +257,6 @@ class ResponseMapAll:
                 continue
             start_ts = record.get(self._start_column)
 
-            if end_ts < start_ts:
-                warn('Record data is invalid. '
-                     'The end time of the path is recorded before the start time.',
-                     UserWarning)
-                continue
-
             if start_ts not in self._start_timestamps:
                 self._start_timestamps.insert(0, start_ts)
                 self._end_timestamps.insert(0, end_ts)
@@ -272,60 +266,6 @@ class ResponseMapAll:
                 if end_ts < self._end_timestamps[idx]:
                     self._end_timestamps[idx] = end_ts
                     self._records[idx] = record
-
-    def to_worst_with_external_latency_case_records(self) -> RecordsInterface:
-
-        end_timestamps: list[int] = []
-        start_timestamps: list[int] = []
-        worst_to_best_timestamps: list[int] = []
-        for start_ts, end_ts, prev_start_ts in zip(self._start_timestamps[1:],
-                                                   self._end_timestamps[1:],
-                                                   self._start_timestamps[:-1]):
-            if end_ts not in end_timestamps:
-                start_timestamps.append(start_ts)
-                end_timestamps.append(end_ts)
-                worst_to_best_timestamps.append(start_ts - prev_start_ts)
-            else:
-                idx = end_timestamps.index(end_ts)
-                if start_ts < start_timestamps[idx]:
-                    start_timestamps[idx] = start_ts
-                    worst_to_best_timestamps[idx] = start_ts - prev_start_ts
-
-        records = self._create_empty_records()
-        for start_ts, end_ts, worst_to_best_ts in sorted(zip(start_timestamps,
-                                                             end_timestamps,
-                                                             worst_to_best_timestamps),
-                                                         key=lambda x: x[0]):
-            record = {
-                self._start_column: start_ts - worst_to_best_ts,
-                'response_time': end_ts - (start_ts - worst_to_best_ts)
-            }
-            records.append(record)
-
-        return records
-
-    def to_best_case_records(self) -> RecordsInterface:
-
-        end_timestamps: list[int] = []
-        start_timestamps: list[int] = []
-        for start_ts, end_ts in zip(self._start_timestamps, self._end_timestamps):
-            if end_ts not in end_timestamps:
-                start_timestamps.append(start_ts)
-                end_timestamps.append(end_ts)
-            else:
-                idx = end_timestamps.index(end_ts)
-                if start_ts > start_timestamps[idx]:
-                    start_timestamps[idx] = start_ts
-
-        records = self._create_empty_records()
-        for start_ts, end_ts in sorted(zip(start_timestamps, end_timestamps), key=lambda x: x[0]):
-            record = {
-                self._start_column: start_ts,
-                'response_time': end_ts - start_ts
-            }
-            records.append(record)
-
-        return records
 
     def to_all_records(self) -> RecordsInterface:
         records = self._create_empty_records()
@@ -352,7 +292,7 @@ class ResponseMapAll:
                     start_timestamps[idx] = start_ts
 
         records = self._create_empty_records()
-        for start_ts, end_ts in sorted(zip(start_timestamps, end_timestamps), key=lambda x: x[0]):
+        for start_ts, end_ts in zip(start_timestamps, end_timestamps):
             record = {
                 self._start_column: start_ts,
                 'response_time': end_ts - start_ts
@@ -567,7 +507,7 @@ class ResponseTime:
             - {'response_time'}
 
         """
-        return self._response_map_all.to_best_case_records()
+        return self._timeseries.to_best_case_records()
 
     def to_worst_with_external_latency_case_records(self) -> RecordsInterface:
         """
@@ -587,7 +527,7 @@ class ResponseTime:
             - {'response_time'}
 
         """
-        return self._response_map_all.to_worst_with_external_latency_case_records()
+        return self._timeseries.to_worst_with_external_latency_case_records()
 
     def to_all_stacked_bar(self) -> RecordsInterface:
         """
