+++ src/test/record/records_service/test_response_time.py
@@ -474,6 +474,88 @@ class TestResponseHistogram:
 
 class TestResponseTimeseries:
 
+    def test_empty_flow_case(self):
+        records_raw = [
+        ]
+        columns = [ColumnValue('start'), ColumnValue('end')]
+
+        records = create_records(records_raw, columns)
+        response = ResponseTime(records)
+
+        response_time = response.to_best_case_records()
+        expect = []
+        assert to_dict(response_time) == expect
+
+        response_time = response.to_worst_with_external_latency_case_records()
+        expect = []
+        assert to_dict(response_time) == expect
+
+    def test_single_flow_case(self):
+        records_raw = [
+            {'start': 0, 'end': 1},
+        ]
+        columns = [ColumnValue('start'), ColumnValue('end')]
+
+        records = create_records(records_raw, columns)
+        response = ResponseTime(records)
+
+        response_time = response.to_best_case_records()
+        expect = []
+        assert to_dict(response_time) == expect
+
+        response_time = response.to_worst_with_external_latency_case_records()
+        expect = []
+        assert to_dict(response_time) == expect
+
+    def test_double_flow_case(self):
+        records_raw = [
+            {'start': 0, 'end': 1},
+            {'start': 2, 'end': 3},
+        ]
+        columns = [ColumnValue('start'), ColumnValue('end')]
+
+        records = create_records(records_raw, columns)
+        response = ResponseTime(records)
+
+        response_time = response.to_best_case_records()
+        expect = [
+            {'start_max': 2, 'response_time': 1}
+        ]
+        assert to_dict(response_time) == expect
+
+        response_time = response.to_worst_with_external_latency_case_records()
+        expect = [
+            {'start_min': 0, 'response_time': 3}
+        ]
+        assert to_dict(response_time) == expect
+
+    # NOTE: Is this test up to specification?
+    def test_cross_flow_case(self):
+        records_raw = [
+            {'start': 0, 'end': 10},
+            {'start': 3, 'end': 4},
+            {'start': 4, 'end': 8},
+            {'start': 6, 'end': 6},
+        ]
+        columns = [ColumnValue('start'), ColumnValue('end')]
+
+        records = create_records(records_raw, columns)
+        response = ResponseTime(records)
+
+        response_time = response.to_best_case_records()
+        expect = [
+            {'start_max': 3, 'response_time': 1},
+            {'start_max': 6, 'response_time': 0}
+        ]
+        assert to_dict(response_time) == expect
+
+        response_time = response.to_worst_with_external_latency_case_records()
+        expect = [
+            {'start_min': 0, 'response_time': 4},
+            {'start_min': 3, 'response_time': 3}
+        ]
+        assert to_dict(response_time) == expect
+
     class TestMultiColumnCase:
         records_raw = [
             {'column0': 5, 'column1': 10, 'column2': 15},  # flow 1 [used as first data]
@@ -510,3 +592,131 @@ class TestResponseTimeseries:
 
             output_dict = to_dict(records)
             assert output_dict == expect
+
+        def test_to_best_case_timeseries(self):
+            response = ResponseTime(
+                create_records(self.records_raw, self.columns),
+                columns=self.column_names
+            )
+            response_time = response.to_best_case_records()
+            expect = [
+                {'column0_max': 20, 'response_time': 10},
+                {'column0_max': 35, 'response_time': 10}
+            ]
+            assert to_dict(response_time) == expect
+
+        def test_to_worst_case_timeseries(self):
+            response = ResponseTime(
+                create_records(self.records_raw, self.columns),
+                columns=self.column_names
+            )
+            response_time = response.to_worst_with_external_latency_case_records()
+            expect = [
+                {'column0_min': 5, 'response_time': 25},
+                {'column0_min': 20, 'response_time': 25}
+            ]
+            assert to_dict(response_time) == expect
+
+
+class TestResponseTimeWorstInInput:
+
+    def test_empty_case(self):
+        records_raw = []
+        columns = [ColumnValue('start'), ColumnValue('end')]
+        records = create_records(records_raw, columns)
+
+        response_time = ResponseTime(records)
+
+        expect_raw = []
+        result = to_dict(response_time.to_worst_case_records())
+        assert result == expect_raw
+
+    def test_two_column_default_case(self):
+        records_raw = [
+            {'start': 0, 'end': 2},
+            {'start': 3, 'end': 4},
+            {'start': 11, 'end': 12}
+        ]
+        columns = [ColumnValue('start'), ColumnValue('end')]
+        records = create_records(records_raw, columns)
+
+        response_time = ResponseTime(records)
+
+        expect_raw = [
+            {'start': 0, 'response_time': 2},
+            {'start': 3, 'response_time': 1},
+            {'start': 11, 'response_time': 1}
+        ]
+        result = to_dict(response_time.to_worst_case_records())
+        assert result == expect_raw
+
+    def test_three_column_default_case(self):
+        records_raw = [
+            {'start': 0, 'middle': 1, 'end': 2},
+            {'start': 3, 'middle': 4, 'end': 6},
+            {'start': 11, 'middle': 13, 'end': 16}
+        ]
+        columns = [ColumnValue('start'), ColumnValue('middle'), ColumnValue('end')]
+        records = create_records(records_raw, columns)
+
+        response_time = ResponseTime(records)
+
+        expect_raw = [
+            {'start': 0, 'response_time': 2},
+            {'start': 3, 'response_time': 3},
+            {'start': 11, 'response_time': 5}
+        ]
+        result = to_dict(response_time.to_worst_case_records())
+        assert result == expect_raw
+
+    def test_single_input_multi_output_case(self):
+        records_raw = [
+            {'start': 0, 'middle': 4, 'end': 5},
+            {'start': 0, 'middle': 4, 'end': 6},
+            {'start': 0, 'middle': 12, 'end': 13}
+        ]
+        columns = [ColumnValue('start'), ColumnValue('middle'), ColumnValue('end')]
+        records = create_records(records_raw, columns)
+
+        response_time = ResponseTime(records)
+
+        expect_raw = [
+            {'start': 0, 'response_time': 5}
+        ]
+        result = to_dict(response_time.to_worst_case_records())
+        assert result == expect_raw
+
+    def test_multi_input_single_output_case(self):
+        records_raw = [
+            {'start': 0, 'middle': 4, 'end': 13},
+            {'start': 1, 'middle': 4, 'end': 13},
+            {'start': 5, 'middle': 12, 'end': 13}
+        ]
+        columns = [ColumnValue('start'), ColumnValue('middle'), ColumnValue('end')]
+        records = create_records(records_raw, columns)
+
+        response_time = ResponseTime(records)
+
+        expect_raw = [
+            {'start': 0, 'response_time': 13}
+        ]
+        result = to_dict(response_time.to_worst_case_records())
+        assert result == expect_raw
+
+    def test_drop_case(self):
+        records_raw = [
+            {'start': 0, 'middle': 4, 'end': 8},
+            {'start': 1, 'middle': 4},
+            {'start': 5, 'middle': 12, 'end': 13}
+        ]
+        columns = [ColumnValue('start'), ColumnValue('middle'), ColumnValue('end')]
+        records = create_records(records_raw, columns)
+
+        response_time = ResponseTime(records)
+
+        expect_raw = [
+            {'start': 0, 'response_time': 8},
+            {'start': 1, 'response_time': 12}
+        ]
+        result = to_dict(response_time.to_worst_case_records())
+        assert result == expect_raw
